style: separate different init func & reuse func
