+++ resources/assets/js/_tests_/model/reducer/post.test.js
@@ -1,13 +1,14 @@
 import cloneDeep from 'lodash/cloneDeep';
 import postReducer from '~/model/reducer/post';
 import { SET_POST_FORM, TOGGLE_STATIS_DATA } from '~/model/action/post';
-import initPost, { initComment, initStudy } from './initPost';
+import initPost from './post';
+import { initComment } from './post/initComment';
+import { initStudy } from './post/initStudy';
 import { transObjToKeysTable } from '~/utils/redux/components/modal/transFormData';
 import result from 'lodash/result';
 
 // init postForm
 const postForm = initPost;
-
 /**
  * global function
  *
